FEAT Initial setup using Nix/home-manager and Lua for Neovim
We got a completely new configuration and, in order to be reproducible,
it is based on the horrendous Nix package manager. The language is not
understandable to me, however it works now and I can add packages and
configurations at least to the home-manager config.
As we did a big purge, I started from scratch, completely switching to
Neovim and its configuration to Lua. What a relief! Never looking back
to vim /VimScript.

These dotfiles are the first version of a working config. It will be
extended step by step.

Next TODOs:
- Add an extensive ReadMe.md
- Configure tmux with keybindings like in QTile using tilish
- Configure zsh
- Add some more goodies to the Neovim-configuration.

What a breeze of fresh air!